feat: add business logic only admins can delete any book

+++ src/main/java/io/github/com/lucasmartinsvieira/trackr/services/BookService.java
@@ -4,6 +4,9 @@ import io.github.com.lucasmartinsvieira.trackr.api.dtos.books.OpenLibrarySeachRe
 import io.github.com.lucasmartinsvieira.trackr.api.dtos.books.OpenLibrarySearchReponseDTO;
 import io.github.com.lucasmartinsvieira.trackr.api.external.OpenLibraryClient;
 import io.github.com.lucasmartinsvieira.trackr.domain.book.Book;
+import io.github.com.lucasmartinsvieira.trackr.domain.user.User;
+import io.github.com.lucasmartinsvieira.trackr.domain.user.UserRole;
+import io.github.com.lucasmartinsvieira.trackr.infra.exception.BookAccessDeniedException;
 import io.github.com.lucasmartinsvieira.trackr.repositories.BookRepository;
 import org.springframework.stereotype.Service;
 
@@ -28,7 +31,13 @@ public class BookService {
         return this.bookRepository.findAll();
     }
 
-    public void deleteBook(UUID id) {
+    public void deleteBook(UUID id, User user) {
+        var book = this.bookRepository.getReferenceById(id);
+
+        if (!book.getUser().equals(user) && user.getRole() != UserRole.ADMIN) {
+            throw new BookAccessDeniedException("This book register does not belong to you");
+        }
+
         this.bookRepository.deleteById(id);
     }
 }
